requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-pubsub==2.29.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
importlib_metadata==8.6.1
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
rsa==4.9.1
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
zipp==3.21.0
